combine some shaderlang files
