--- src/graphics/shaderlang/func.h
@@ -1,9 +1,0 @@
-#pragma once
-
-#include <string>
-
-namespace ramiel::shaderlang {
-
-    std::string readShaderFunc(const std::string& header);
-
-}
