--- src/graphics/shaderlang/pixelshader.h
@@ -1,10 +1,0 @@
-#pragma once
-
-#include <string>
-#include "template.h"
-
-namespace ramiel::shaderlang {
-
-    TemplateArgs compilePixelShader();
-
-}
