+++ src/graphics/shaderlang/read.cpp
@@ -1,5 +1,5 @@
 #include <iostream>
-#include "attribute.h"
+#include "read.h"
 #include "token.h"
 
 namespace {
@@ -27,6 +27,11 @@ namespace {
 
     bool isValidAttrName(const std::string& token) {
         return std::isalpha(token[0]) || token[0] == '_';
+    }
+    
+
+    bool isPunctuation(const std::string& token) {
+        return token.length() == 1 && !std::isalnum(token[0]) && token[0] != '_';
     }
 
 }
@@ -89,6 +94,36 @@ namespace ramiel::shaderlang {
         exit(1);
 
         return attrs;
+    }
+
+
+    std::string readShaderFunc(const std::string& header) {
+        auto& headerToken = popToken();
+        if (headerToken != header) {
+            std::cout << "error: expected \"" << header << "\", got \"" << headerToken << "\"\n";
+            exit(1);
+        }
+
+        auto& brace = popToken();
+        if (brace != "{") {
+            std::cout << "error: expected \"{\", got \"" << brace << "\"\n";
+            exit(1);
+        }
+
+        std::string shaderFunc = "{";
+        size_t braceDepth = 1;
+        
+        while (braceDepth) {
+            auto& token = popToken();
+            if (token == "{") braceDepth++;
+            else if (token == "}") braceDepth--;
+            shaderFunc += token;
+
+            // this is a freaky workaround until we stop relying on c++
+            if (!isPunctuation(token)) shaderFunc += ' ';
+        }
+
+        return shaderFunc;
     }
 
 }
