+++ src/graphics/shaderlang/shader.cpp
@@ -1,10 +1,9 @@
 #include <iostream>
 #include <vector>
 
-#include "vertexshader.h"
+#include "shader.h"
 #include "token.h"
-#include "attribute.h"
-#include "func.h"
+#include "read.h"
 #include "stringify.h"
 
 namespace ramiel::shaderlang {
@@ -39,6 +38,30 @@ namespace ramiel::shaderlang {
             { "ATTR_IN_DEF", stringifyAttrInDef(inAttrs) },
             { "ATTR_OUT_DEF", stringifyAttrOutDef(outAttrs) },
             { "ATTR_IN_INTERPRET", stringifyAttrInInterpret(inAttrs) },
+            { "FUNC", shaderFunc }
+        };
+
+        return templateArgs;
+    }
+
+
+    TemplateArgs compilePixelShader() {
+        auto& shaderName = popToken();
+        std::vector<Attr> stateAttrs = readAttrList("STATE");
+        std::vector<Attr> inAttrs = readAttrList("IN");
+
+        std::string initFunc = readShaderFunc("INIT");
+        std::string shaderFunc = readShaderFunc("SHADER");
+
+        TemplateArgs templateArgs = {
+            { "NAME", shaderName },
+            { "STATE", stringifyState(stateAttrs) },
+            { "CTOR_ARGS", stringifyCtorArgs(stateAttrs) },
+            { "CTOR_INIT", stringifyCtorInit(stateAttrs) },
+            { "ATTR_IN_TYPE", stringifyAttrType(inAttrs) },
+            { "ATTR_IN_DEF", stringifyAttrInDef(inAttrs) },
+            { "ATTR_IN_INTERPRET", stringifyAttrInInterpret(inAttrs) },
+            { "FUNC_INIT", initFunc },
             { "FUNC", shaderFunc }
         };
