+++ src/graphics/shaderlang/shader.h
@@ -1,10 +1,10 @@
 #pragma once
 
-#include <string>
 #include "template.h"
 
 namespace ramiel::shaderlang {
 
     TemplateArgs compileVertexShader();
+    TemplateArgs compilePixelShader();
 
 }
